requirements.txt
streamlit==1.41.1
langchain==0.3.13
langchain-community==0.3.13
langchain-core==0.3.28
langchain-openai==0.2.14
langchain-ollama==0.2.0
langchain-text-splitters==0.3.4
langchainhub==0.1.21
openai==1.58.1
tavily-python==0.7.12
python-dotenv==1.2.1
requests==2.32.3
pypdf==5.1.0
fastembed==0.4.2
pydantic==2.10.6
SQLAlchemy==2.0.36
validators==0.34.0
docarray>=0.32.1
